Some small cleanups and movement i made while i was borde.
