+++ packages/kartoffelgames.web.potato_web_builder/test/utility/test-dom-environment.ts
@@ -46,65 +46,3 @@ export default function globalJsdom(html = defaultHtml, options = {}) {
 }
 
 globalJsdom();
-
-/**
- * // TODO: For later.
- */
-/*
-const globalDefaultTarget = (pGlobalThis: DOMWindow): InteractionZoneGlobalScopeTarget => {
-	// Create default globalThis target.
-	const lTarget = {
-		target: pGlobalThis,
-		patches: {
-			promise: pGlobalThis.Promise?.name,
-			eventTarget: pGlobalThis.EventTarget?.name,
-			classes: new Array<string>(),
-			functions: new Array<string>()
-		}
-	} satisfies InteractionZoneGlobalScopeTarget;
-
-	// Add all asyncron functions.
-	const lAsyncFunctionNames: Array<string | undefined> = [
-		pGlobalThis.requestAnimationFrame?.name,
-		pGlobalThis.setInterval?.name,
-		pGlobalThis.setTimeout?.name
-	];
-	lTarget.patches.functions.push(...lAsyncFunctionNames.filter(pClass => !!pClass) as Array<string>);
-
-	// Add all global classes with events.
-	const lDomClassNames: Array<string | undefined> = [
-		pGlobalThis.XMLHttpRequestEventTarget?.name,
-		pGlobalThis.XMLHttpRequest?.name,
-		pGlobalThis.Document?.name,
-		pGlobalThis.SVGElement?.name,
-		pGlobalThis.Element?.name,
-		pGlobalThis.HTMLElement?.name,
-		pGlobalThis.HTMLMediaElement?.name,
-		pGlobalThis.HTMLFrameSetElement?.name,
-		pGlobalThis.HTMLBodyElement?.name,
-		pGlobalThis.HTMLFrameElement?.name,
-		pGlobalThis.HTMLIFrameElement?.name,
-		pGlobalThis.HTMLMarqueeElement?.name,
-		pGlobalThis.Worker?.name,
-		pGlobalThis.IDBRequest?.name,
-		pGlobalThis.IDBOpenDBRequest?.name,
-		pGlobalThis.IDBDatabase?.name,
-		pGlobalThis.IDBTransaction?.name,
-		pGlobalThis.WebSocket?.name,
-		pGlobalThis.FileReader?.name,
-		pGlobalThis.Notification?.name,
-		pGlobalThis.RTCPeerConnection?.name
-	];
-	lTarget.patches.classes.push(...lDomClassNames.filter(pClass => !!pClass) as Array<string>);
-
-	// Add all global classes with async callbacks.
-	const lObserverClassNames: Array<string | undefined> = [
-		pGlobalThis.ResizeObserver?.name,
-		pGlobalThis.MutationObserver?.name,
-		pGlobalThis.IntersectionObserver?.name
-	];
-	lTarget.patches.classes.push(...lObserverClassNames.filter(pClass => !!pClass) as Array<string>);
-
-	return lTarget;
-};
-*/
